Revert "boxing and unboxing"
This reverts commit 5a24ad5c

+++ src/test/java/org/apache/commons/lang3/ArrayUtilsRemoveTest.java
@@ -724,7 +724,7 @@ public class ArrayUtilsRemoveTest {
 
     @Test
     public void testRemoveNumberArray() {
-        final Number[] inarray = {1, 2L, (byte) 3};
+        final Number[] inarray = {Integer.valueOf(1), Long.valueOf(2), Byte.valueOf((byte) 3)};
         assertEquals(3, inarray.length);
         Number[] outarray;
         outarray = ArrayUtils.remove(inarray, 1);
